Added SSL transport support.
Added support for IPC over SSL. The server can load an X509Certificate2 object as the certificate, and the client can either rely upon the system certificate validation or implement a custom validation callback.

While the code doesn't currently implement client certificates, it was written with the expectation that this functionality could be added later.