SSL and stream translation support on TCP
The previous approach to implementing SSL was to use separate host and client classes for SSL, but this ended up duplicating a lot of code. This has now been rewritten to make SSL an option on the TCP side. This commit also adds support for stream translators, which can be used to add custom headers and other traffic modifications.